Fix cluster species data datatypes

+++ UI/src/mainwindow/cluster.cpp
@@ -156,11 +156,11 @@ std::unordered_map<int, int> Cluster::speciesSizes() {
   return species_sizes;
 }
 
-std::vector<std::tuple<double, double, double>> Cluster::getCurrentSpeciesData() {
+std::vector<std::tuple<int, double, int>> Cluster::getCurrentSpeciesData() {
   std::lock_guard<std::recursive_mutex> lock(mutex_);
 
   std::unordered_map<int, int> species_sizes = this->speciesSizes();
-  std::vector<std::tuple<double, double, double>> species_data;
+  std::vector<std::tuple<int, double, int>> species_data;
 
   for (const auto& pair : species_sizes) {
     species_data.push_back(
@@ -242,7 +242,7 @@ void Cluster::update_creatures_species(
   }
 }
 
-std::vector<std::tuple<double, double, double>> Cluster::getSpeciesData() {
+std::vector<std::tuple<int, double, int>> Cluster::getSpeciesData() {
   std::lock_guard<std::recursive_mutex> lock(mutex_);
 
   return species_data_;
